% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document.R
\docType{data}
\name{b_cell_markers}
\alias{b_cell_markers}
\title{B-cell marker genes}
\format{
A character vector with 50 entries:
\describe{
  \item{CD19}{Pan-B-cell marker}
  \item{CD20}{B-cell activation marker}
  \item{CD22}{B-cell receptor signaling}
}
}
\source{
https://doi.org/10.1186/s13059-016-1070-5
}
\usage{
b_cell_markers
}
\description{
A character vector of genes marking B-cell subsets and functional states.
}
\keyword{datasets}
