% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrepareTreeMap.R
\name{DataInit}
\alias{DataInit}
\title{Extract a slice of point cloud around DBH height for TreeMap}
\usage{
DataInit(Mydata, ThickDHP = 0.1, Hsoil = 4)
}
\arguments{
\item{Mydata}{this dataset should contain at least 3 columns with X, Y, Z values.
The .las (.laz) format is better for faster processing but a data.frame (or data.table) is acceptable.}

\item{ThickDHP}{indicate the thickness of the slice around the DBH (1.3m height) (10cm by default, so the thick of the slice will be 20cm)}

\item{Hsoil}{Needed for the function Zsoil. Indicates the height of the points held above ground to create the DTM.
If the field is generally flat, the 4 m by default are sufficient, however for a fairly steep terrain,
it may be necessary to increase the value of Zsoil.}
}
\value{
Returns a data.table containing the points X, Y, Z, and Zr (created with Zsoil) located between (1.3 - ThickDHP) & (1.3 + ThickDHP)
}
\description{
Extract a slice centered on the DBH height (1.3m) with a thickness of 2* ThickDHP
}
