% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document.R
\docType{data}
\name{endothelial}
\alias{endothelial}
\title{Endothelial markers}
\format{
A character vector with 5 entries:
\describe{
  \item{CD31}{Endothelial cell adhesion molecule}
  \item{VEGFR}{Vascular endothelial growth factor receptor}
}
}
\source{
https://doi.org/10.1186/s13059-016-1070-5
}
\usage{
endothelial
}
\description{
A character vector of genes specific to endothelial cells.
}
\keyword{datasets}
