% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document.R
\docType{data}
\name{fibroblast}
\alias{fibroblast}
\title{Fibroblast markers}
\format{
A character vector with 10 entries:
\describe{
  \item{FAP}{Fibroblast activation protein}
  \item{COL1A1}{Collagen type I alpha 1 chain}
}
}
\source{
https://doi.org/10.1186/s13059-016-1070-5
}
\usage{
fibroblast
}
\description{
A character vector of genes specific to fibroblast cells.
}
\keyword{datasets}
