% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document.R
\docType{data}
\name{follicular_dc}
\alias{follicular_dc}
\title{Follicular Dendritic Cell (FDC) Markers}
\format{
A character vector with 8 entries:
\describe{
  \item{CD21 (CR2)}{Complement receptor}
  \item{CD23 (FCER2)}{IgE receptor}
  \item{CXCL13}{Chemokine for B-cell follicle organization}
}
}
\source{
From DeepSeek
}
\usage{
follicular_dc
}
\description{
A character vector of genes specific to follicular dendritic cells.
}
\keyword{datasets}
