% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document.R
\docType{data}
\name{g2m.genes}
\alias{g2m.genes}
\title{Cell cycle marker genes (G2/M phase)}
\format{
A character vector.
}
\source{
Processed from [original source, e.g., "Tirosh et al. 2016, Science"]
}
\usage{
g2m.genes
}
\description{
A character vector of genes associated with the G2/M phase of the cell cycle.
}
\keyword{datasets}
