% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document.R
\docType{data}
\name{myeloid_dc}
\alias{myeloid_dc}
\title{Myeloid Dendritic Cell Markers}
\format{
A character vector with 6 entries:
\describe{
  \item{CD1A/CD1B}{Antigen presentation}
  \item{CLEC10A}{cDC2 subset marker}
}
}
\source{
https://doi.org/10.1186/s13059-016-1070-5
}
\usage{
myeloid_dc
}
\description{
A character vector of genes marking myeloid dendritic cell subsets.
}
\keyword{datasets}
