% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document.R
\docType{data}
\name{myeloid_markers}
\alias{myeloid_markers}
\title{Myeloid markers}
\format{
A character vector with 15 entries:
\describe{
  \item{CD14}{Monocyte marker}
  \item{CD68}{Macrophage marker}
}
}
\source{
https://doi.org/10.1186/s13059-016-1070-5
}
\usage{
myeloid_markers
}
\description{
A character vector of genes marking myeloid cell types.
}
\keyword{datasets}
