% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document.R
\docType{data}
\name{nk_cell_markers}
\alias{nk_cell_markers}
\title{NK cell marker genes}
\format{
A character vector with 12 entries:
\describe{
  \item{NKG7}{NK cell activation marker}
  \item{CD56}{NK cell subset marker}
}
}
\source{
https://doi.org/10.1186/s13059-016-1070-5
}
\usage{
nk_cell_markers
}
\description{
A character vector of genes marking natural killer (NK) cell subsets.
}
\keyword{datasets}
