% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document.R
\docType{data}
\name{s.genes}
\alias{s.genes}
\title{Cell cycle marker genes (S phase)}
\format{
A character vector.
}
\source{
Processed from [original source, e.g., "Tirosh et al. 2016, Science"]
}
\usage{
s.genes
}
\description{
A character vector of genes associated with the S phase of the cell cycle.
}
\keyword{datasets}
