% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document.R
\docType{data}
\name{t_cell_markers}
\alias{t_cell_markers}
\title{T-cell marker genes}
\format{
A character vector with 45 entries:
\describe{
  \item{CD3D/CD3E/CD3G}{Pan-T-cell markers}
  \item{CD4}{Helper T-cells}
  \item{CD8A/CD8B}{Cytotoxic T-cells}
  \item{FOXP3}{Regulatory T-cells (Tregs)}
  \item{PDCD1}{Exhaustion marker}
}
}
\source{
https://doi.org/10.1186/s13059-016-1070-5
}
\usage{
t_cell_markers
}
\description{
A character vector of genes marking T-cell subsets and functional states.
}
\keyword{datasets}
