% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrepareTreeMap.R
\name{testAlphaThreshold}
\alias{testAlphaThreshold}
\title{Find the best threshold (Alpha)}
\usage{
testAlphaThreshold(data, standName = NULL, Threshold = c(3, 5, 8, 10),
  saveGraphPath = "./TLSAnalysis")
}
\arguments{
\item{data}{must be the data generated by the function DataInit to work}

\item{standName}{Optional. If the tested plot has a name, it will be added to the name of the graphics (name = TestStd by default).}

\item{Threshold}{Indicates the threshold value to be tested (can be a vector of several values). A look at the graphs generated will allow to adjust the desired values and retest them.}

\item{saveGraphPath}{indicates the folder where the graphs will be generated.}
}
\value{
Returns graphs stored in the chosen folder
}
\description{
Generated graphs for different threshold values to select the best value for cleaning the point cloud and make it usable for the TreeMap.
The selected final threshold value should be used to remove single points and branch points. Only isolated trunks should remain visible
}
