% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document.R
\docType{data}
\name{tls_50_genes}
\alias{tls_50_genes}
\title{TLS Signature Genes}
\format{
A character vector.
}
\source{
Processed from [your source, e.g., "public dataset XYZ"]
}
\usage{
tls_50_genes
}
\description{
A character vector of 50 genes associated with tertiary lymphoid structures (TLS).
}
\keyword{datasets}
