% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tls-ls.R
\name{TLS_Gen}
\alias{TLS_Gen}
\title{Calculate TLS or OLS solution}
\usage{
TLS_Gen(In, y, mode = 1, scaling = 0, ef = 0)
}
\arguments{
\item{In}{- equations, In(n,m)}

\item{y}{- observations, y(n)}

\item{mode}{- type of solution, 1 - SVD, 2 - eigen values}

\item{scaling}{- sort of scaling in case of TLS solution, bit mask: 1 - by rows, 2 - by columns, 3 - Frobenian}

\item{ef}{- error-free variables qty}
}
\value{
List of variables, where  
  X - solution, X(m)
  E - discripency, E(n)
  S - Singular Values or eigen values, depends on mode, S(m+1)
  Corr - covariance matrix or corrilation matrix, Corr(m,m)
  s0 - mean square deviation of the model
  s0_dis - mean square deviation of the solution by discripency 
  s_X - solution errors, s_X(m)
  err - error code or condition number
}
\description{
Calculate TLS or OLS solution
}
