% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TreeMap.R
\name{TreeMap_Auto}
\alias{TreeMap_Auto}
\title{Automatic creation of a TreeMap}
\usage{
TreeMap_Auto(Mydata, standName = NULL, alpha, minPts = 8, eps = 0.1,
  minDbh = 0, savePlot = T, TreeValue = NULL,
  savePath = "./TLSAnalysis/")
}
\arguments{
\item{Mydata}{must be the data generated by the function DataInit to work}

\item{standName}{Optional. If the tested plot has a name, it will be added to the name of the graphics (name = TestStd by default).}

\item{alpha}{Alpha is the threshold value needed to clean the point cloud data}

\item{minPts}{Minimum distance between 2 points to belong to the same cluster}

\item{eps}{Minimal spacing between 2 points to create 2 clusters}

\item{minDbh}{Minimum DBH value saved (if the value is greater than 0, trees with a lower DBH will be removed from the TreeMap)}

\item{savePlot}{Saves a map of the position and DBH of all trees (This map will be used during the manual step to possibly detect adjustment errors or missing trees)}

\item{TreeValue}{Used to add particular trees to the map}

\item{savePath}{indicates the folder where the map and the data will be saved}
}
\value{
Generate a .txt file containing the position and the DBH of all the trees as well as a map (Treedetection.jpg) positioning them on the plot
}
\description{
Detect the coordinates XY and the Diameter at breast height (DBH) of all trees on the plot.
The point cloud had to be prepared and sliced with the function 'DataInit' and a value of threshold had to be found (use testAlphaThreshold())
}
