% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TreeMap.R
\name{TreeMap_semiManual}
\alias{TreeMap_semiManual}
\title{Semi-automatic TreeMap corrections}
\usage{
TreeMap_semiManual(Mydata, standName = NULL, alpha, critR = 0.1,
  Dist_view = 2, add3Dhelp = F, savePath = "./TLSAnalysis/",
  TreeValue = NULL)
}
\arguments{
\item{Mydata}{must be the data generated by the function DataInit to work}

\item{standName}{Optional. If the tested plot has a name, it will be added to the name of the graphics (name = TestStd by default).
WARNING! If you work with multiple plots and do not specify a unique name for each new plot, the default name will be used each time and the TreeMap of the previous plot will be overwritten.}

\item{alpha}{Alpha is the threshold value needed to clean the point cloud data}

\item{critR}{Indicates the threshold value of the variable 'critR' for which a tree will be displayed and will require a correction}

\item{Dist_view}{Indicates the distance displayed around the target tree (Dist_view x DBH)}

\item{add3Dhelp}{if TRUE, a 3d plot of the tree will be displyed to help correction}

\item{savePath}{indicates the folder where the map and the data will be saved}

\item{TreeValue}{Used to add particular trees to the map}
}
\value{
Generate a .txt file containing the position and the DBH of all the trees as well as a map (Treedetection.jpg) positioning them on the plot
}
\description{
From the TreeMap created with the function 'TreeMap_Auto', all trees having a too high value
of the variable 'critR' will be considered as potentially problematic and will be displayed to be manually corrected.
To correct the displayed tree, please click on the center of the tree, then click on Esc.
It is possible to click up to 5 times if necessary (i.e. 5 trees need to be positionned on the same image).
In this case, click on Esc only at the end of the multiple selection.
}
