% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Basics.R
\name{ValidLibrary}
\alias{ValidLibrary}
\title{Check and install packages}
\usage{
ValidLibrary(x)
}
\arguments{
\item{x}{x is a list of packages}
}
\description{
This function tests whether the packages in the x list are loaded. It tests if they are installed too. If not, they will be installed
}
