% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sol_Extract.R
\name{Zsoil}
\alias{Zsoil}
\title{Zsoil : recalculates Z values}
\usage{
Zsoil(Mydata, Hsoil = 4)
}
\arguments{
\item{Mydata}{This dataset should contain at least 3 columns with X, Y, Z values.
The .las (.laz) format is better for faster processing but a data.frame (or data.table) is acceptable.}

\item{Hsoil}{Indicates the height of the points held above ground to create the DTM.
If the field is generally flat, the 4 m by default are sufficient, however for a fairly steep terrain,
it may be necessary to increase the value of Zsoil.}
}
\value{
Returns a file in the las format containing the columns X, Y and Z as well as the new column Zr.
}
\description{
This function recalculates the Z value based on the distance to the ground.
}
