% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runKNN.R
\name{runKNN}
\alias{runKNN}
\title{Run K-Nearest Neighbors for TLS Identification}
\usage{
runKNN(seurat, exp_threshold = 0.7, min_spots = 3, max_distance = 8)
}
\arguments{
\item{seurat}{A \code{Seurat} object containing spatial transcriptomics data.}

\item{exp_threshold}{A numeric value (default: 0.70) indicating the expression threshold
for identifying high-expression spots.}

\item{min_spots}{An integer (default: 3) representing the minimum number of adjacent spots
required to form a valid TLS cluster.}

\item{max_distance}{A numeric value (default: 8) defining the maximum distance between
adjacent spots to be considered neighbors.}
}
\value{
A modified \code{Seurat} object with an additional metadata column \code{TLS_identity}
        indicating whether each spot is part of a TLS or not.
}
\description{
Identifies tumor-associated lymphoid structures (TLS) in spatial transcriptomics data
using a K-nearest neighbors approach based on expression thresholds and spatial proximity.
}
\details{
This function processes each sample in the provided \code{Seurat} object to:
\enumerate{
  \item Subset data to individual samples.
  \item Identify high expression spots based on the specified threshold.
  \item Calculate dynamic neighbors using a K-nearest neighbors approach.
  \item Create an adjacency matrix for identified neighbors.
  \item Find connected components in the adjacency matrix to identify valid TLS clusters.
  \item Label spots as "TLS" or "not TLS" based on their cluster membership.
}
}
\examples{
\donttest{
seurat_obj <- runKNN(seurat, exp_threshold = 0.70, min_spots = 3, max_distance = 8)
}

}
